app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "langchain[openai]>=0.3.17",
    "langchain-community>=0.3.16",
    "langchain-ollama>=0.2.3",
    "langgraph>=0.2.70",
    "pre-commit>=4.1.0",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.5",
    "langsmith>=0.3.8",
    "beautifulsoup4>=4.13.3",
    "langchain-text-splitters>=0.3.6",
    "langchainhub>=0.1.21",
    "tiktoken>=0.9.0",
    "pytest>=8.3.4",
    "langchain-openai>=0.3.6",
    "PyMuPDF>=1.25.3",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "opencv-python>=4.11.0.86",
    "numpy>=2.2.3",
    "fpdf>=1.7.2",
    "docx>=0.2.4",
    "langgraph-checkpoint-sqlite>=2.0.7",
    "aiosqlite>=0.21.0",
]
